Create find bar addon
