+++ addons-l10n/en/editor-devtools.json
@@ -30,11 +30,6 @@
   "editor-devtools/unused-var": "{count, plural, one {Developer tools: Delete 1 unused local variable? Here it is:\n} other {Developer tools: Delete # unused local variables? Here they are:\n} }",
   "editor-devtools/unused-list": "{count, plural, one {Developer tools: Delete 1 unused local list? Here it is:\n} other {Developer tools: Delete # unused local lists? Here they are:\n} }",
   "editor-devtools/var-not-exist": "That variable does not exist...",
-  "editor-devtools/find": "Find",
-  "editor-devtools/find-placeholder": "Find (Ctrl+F)",
-  "editor-devtools/deep": "Deep",
-  "editor-devtools/help": "Help",
-  "editor-devtools/tutorials": "Tutorials",
   "editor-devtools/replace": "Developer tools: Switch all {name} in this sprite for the variable named:",
   "editor-devtools/top": "send to top",
   "editor-devtools/bottom": "send to bottom",
