+++ addons-l10n/en/find-bar.json
@@ -1,0 +1,6 @@
+{
+  "find-bar/find": "Find",
+  "find-bar/find-placeholder": "Find (Ctrl+F)",
+  "find-bar/help": "Help",
+  "find-bar/tutorials": "Tutorials"
+}
