+++ addons/addons.json
@@ -122,6 +122,7 @@
   "default-costume-editor-color",
   "hide-stage",
   "one-click-studio-invites",
+  "find-bar",
 
   "// NEW ADDONS ABOVE THIS ↑↑",
   "// Note: these themes need this exact order to work properly,",
