+++ addons/find-bar/blockly/BlockFlasher.js
