+++ addons/find-bar/blockly/BlockInstance.js
