+++ addons/find-bar/blockly/BlockItem.js
