+++ addons/find-bar/userstyle.css
@@ -1,0 +1,366 @@
+.find-bar {
+  display: flex;
+  align-items: center;
+  white-space: nowrap;
+  font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
+  width: 100%;
+  height: 100%;
+}
+
+.find-bar > label {
+  /* padding instead of margin so clicking on the empty area will select the input */
+  padding-left: 1.5em;
+  padding-right: 1em;
+  font-weight: bold;
+  font-size: 0.625rem;
+  user-select: none;
+  cursor: default;
+  white-space: nowrap;
+  padding-top: 2px;
+}
+
+[dir="rtl"] .find-bar > label {
+  padding-right: 1.5em;
+  padding-left: 1em;
+}
+
+.find-help {
+  transition: opacity 200ms ease-in;
+  margin-left: 0;
+  font-size: 10px;
+  opacity: 0;
+}
+
+.find-bar:hover .find-help {
+  opacity: 0.7;
+}
+
+.find-wrapper {
+  overflow: visible;
+  position: relative;
+  height: 2rem;
+  width: 100%;
+  max-width: 16em;
+}
+
+.find-dropdown-out {
+  display: block;
+  top: -6px;
+  z-index: 100;
+  width: 100%;
+  max-width: 16em;
+  position: relative;
+  padding: 4px;
+  border: none;
+  border-radius: 4px;
+  margin-top: 6px;
+}
+
+.find-dropdown-out.visible {
+  position: absolute;
+  width: 16em;
+  box-shadow: 0px 0px 8px 1px rgba(0, 0, 0, 0.3);
+  background-color: white;
+}
+
+/* We need to modify Scratch styles so that the place where the find bar is injected */
+/* has actually correct size information, which is used to make the find bar not cover up controls */
+[class*="gui_tab-list_"] {
+  width: 100%;
+}
+[class*="gui_tab_"] {
+  flex-grow: 0;
+}
+
+.find-input {
+  width: 100%;
+  box-sizing: border-box !important;
+  /* !important required for extension, because CSS injection method (and hence order) differs from addon */
+  height: 1.5rem;
+
+  /* Change Scratch default styles */
+  border-radius: 0.25rem;
+  font-size: 0.75rem;
+  padding-left: 0.4em;
+}
+
+.find-input:focus {
+  /* Change Scratch default styles */
+  box-shadow: none;
+}
+
+.find-dropdown {
+  display: none;
+  position: relative;
+  padding: 0.2em 0;
+  font-size: 0.75rem;
+  line-height: 1;
+  overflow-y: auto;
+  min-height: 128px;
+  max-height: 65vh;
+  user-select: none;
+  max-width: 100%;
+  margin-top: 6px;
+  border: none;
+}
+
+.find-dropdown-out.visible > .find-dropdown {
+  display: block;
+}
+
+.find-dropdown > li {
+  display: block;
+  padding: 0.5em 0.3em;
+  white-space: nowrap;
+  margin: 0;
+  font-weight: bold;
+  text-overflow: ellipsis;
+  overflow: hidden;
+}
+
+.find-dropdown > li > b {
+  background-color: #aaffaa;
+  color: black;
+}
+
+/* Drop down items */
+.find-dropdown > li:hover,
+.find-dropdown > li.sel {
+  color: white;
+  cursor: pointer;
+}
+
+.find-dropdown > li::before {
+  content: "\25CF "; /* ● */
+}
+
+.define {
+  color: rgb(255, 102, 128);
+}
+.define:hover,
+.define.sel {
+  background-color: rgb(255, 102, 128);
+}
+
+.flag {
+  color: #4cbf56;
+}
+.flag:hover,
+.flag.sel {
+  background-color: #4cbf56;
+}
+
+.receive {
+  color: rgb(255, 191, 0);
+}
+.receive:hover,
+.receive.sel {
+  background-color: rgb(255, 191, 0);
+}
+
+.event {
+  color: rgb(255, 191, 0);
+}
+.event:hover,
+.event.sel {
+  background-color: rgb(255, 191, 0);
+}
+
+.var,
+.VAR {
+  color: rgb(255, 120, 26);
+}
+.var:hover,
+.var.sel,
+.VAR:hover,
+.VAR.sel {
+  background-color: rgb(255, 120, 26);
+}
+
+.list,
+.LIST {
+  color: rgb(255, 102, 26);
+}
+.list:hover,
+.list.sel,
+.LIST:hover,
+.LIST.sel {
+  background-color: rgb(255, 102, 26);
+}
+
+.costume {
+  color: rgb(26, 66, 255);
+}
+.costume:hover,
+.costume.sel {
+  background-color: rgb(26, 66, 255);
+}
+
+.operators {
+  background-color: rgb(64, 191, 74);
+  color: rgb(56, 148, 56);
+}
+.operators:hover,
+.operators.sel {
+  background-color: rgb(56, 148, 56);
+}
+
+.data {
+  background-color: rgb(255, 140, 26);
+  color: rgb(219, 110, 0);
+}
+.data:hover,
+.data.sel {
+  background-color: rgb(219, 110, 0);
+}
+
+.data-lists {
+  background-color: rgb(255, 102, 26);
+  color: rgb(255, 102, 26);
+}
+.data-lists:hover,
+.data-lists.sel {
+  background-color: rgb(207, 81, 20);
+}
+
+.sensing {
+  background-color: rgb(76, 191, 230);
+  color: rgb(46, 142, 184);
+}
+.sensing:hover,
+.sensing.sel {
+  background-color: rgb(46, 142, 184);
+}
+
+.looks {
+  background-color: rgb(153, 102, 255);
+  color: rgb(119, 77, 203);
+}
+.looks:hover,
+.looks.sel {
+  background-color: rgb(119, 77, 203);
+}
+
+.sounds {
+  background-color: rgb(214, 92, 214);
+  color: rgb(189, 66, 189);
+}
+.sounds:hover,
+.sounds.sel {
+  background-color: rgb(189, 66, 189);
+}
+
+.motion {
+  background-color: rgb(76, 151, 255);
+  color: rgb(51, 115, 204);
+}
+.motion:hover,
+.motion.sel {
+  background-color: rgb(51, 115, 204);
+}
+
+.events {
+  background-color: rgb(255, 213, 0);
+  color: rgb(204, 153, 0);
+}
+.events:hover,
+.events.sel {
+  background-color: rgb(204, 153, 0);
+}
+
+.control {
+  background-color: rgb(255, 171, 25);
+  color: rgb(207, 139, 23);
+}
+.control:hover,
+.control.sel {
+  background-color: rgb(207, 139, 23);
+}
+
+.null {
+  background-color: rgb(255, 102, 128);
+  color: rgb(255, 77, 106);
+}
+.null:hover,
+.null.sel {
+  background-color: rgb(255, 77, 106);
+}
+
+.extension {
+  background-color: rgb(15 189 140);
+}
+.extension:hover,
+.extension.sel {
+  background-color: rgb(11 158 117);
+}
+
+/*.find-dropdown>li.hat::before {                           content: '◠ ';}*/
+/*.find-dropdown>li.hat::before {                           content: '▲ ';}*/
+/*.find-dropdown>li.block::before {                         content: '■ ';}*/
+/*.find-dropdown>li.reporter::before {                      content: '● ';}*/
+/*.find-dropdown>li.boolean::before {                       content: '◆ ';}*/
+
+#s3devIDD > li {
+  height: 19px;
+  padding: 3px 8px;
+  margin: 2px 0.3em;
+  box-sizing: border-box;
+  position: relative;
+  color: white;
+  font-weight: bold;
+  width: min-content;
+}
+
+.find-dropdown > li.hat {
+  border-radius: 14px 14px 3px 3px;
+}
+.find-dropdown > li.block {
+  border-radius: 3px;
+}
+.find-dropdown > li.reporter {
+  border-radius: 10px;
+}
+
+.find-dropdown > li.boolean {
+  width: min-content;
+}
+.find-dropdown > li.boolean::before {
+  content: "";
+  position: absolute;
+  left: 0;
+  top: 0;
+  width: 0;
+  height: 0;
+  border-right: 9px solid transparent;
+  border-top: 9px solid white;
+  border-bottom: 10px solid white;
+}
+.find-dropdown > li.boolean::after {
+  content: "";
+  position: absolute;
+  right: 0;
+  top: 0;
+  width: 0;
+  height: 0;
+  border-left: 9px solid transparent;
+  border-top: 9px solid white;
+  border-bottom: 10px solid white;
+}
+
+.find-carousel {
+  font-weight: normal;
+  position: absolute;
+  right: 0;
+  white-space: nowrap;
+  background-color: inherit;
+  z-index: 1;
+  padding: 0;
+}
+
+.find-carousel-control {
+  padding: 0 6px;
+}
+
+.find-carousel-control:hover {
+  color: #ffff80;
+}
